Round1A_PDF_Outlines/requirements.txt
pymupdf==1.24.9
numpy==1.26.4
pyyaml==6.0.2
rapidfuzz==3.9.7
jsonschema==4.23.0
